+++ Project.toml
@@ -17,7 +17,8 @@ PProf = "2, 3"
 julia = "1.6"
 
 [extras]
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["ReTestItems", "Test"]
